requirements.txt
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
numpy==1.20.1
pandas==1.2.3
virtualenv==20.4.2
django-heroku==0.3.1
gunicorn==20.1.0